backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
anytree==2.12.1
ascii_colors==0.5.2
asgiref==3.8.1
attrs==25.3.0
autograd==1.7.0
beartype==0.18.5
beautifulsoup4==4.13.3
black==24.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
Django==5.1.7
django-cors-headers==4.7.0
django-jalali==7.4.0
djangorestframework==3.15.2
filelock==3.18.0
filetype==1.2.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
future==1.0.0
gensim==4.3.3
graspologic==3.4.1
graspologic-native==1.2.3
greenlet==3.1.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.29.3
hyppo==0.4.0
idna==3.10
importlib_metadata==8.6.1
isort==6.0.1
jalali_core==1.0.0
jdatetime==5.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lightrag-hku==1.2.6
litellm==1.60.0
llama-cloud==0.1.15
llama-cloud-services==0.6.6
llama-index==0.12.25
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.25
llama-index-embeddings-huggingface==0.5.2
llama-index-embeddings-litellm==0.3.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.9
llama-index-llms-litellm==0.3.0
llama-index-llms-llama-cpp==0.4.0
llama-index-llms-ollama==0.5.3
llama-index-llms-openai==0.3.26
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.6
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.4.post1
llama_cpp_python==0.3.8
llvmlite==0.44.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
nano-vectordb==0.0.4.3
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==1.26.4
ollama==0.4.7
openai==1.58.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
patsy==1.0.1
pillow==11.1.0
pipmaster==0.5.1
platformdirs==4.3.6
POT==0.9.5
propcache==0.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pynndescent==0.5.13
pyparsing==3.2.1
pypdf==5.4.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.23.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.13.2
selenium==4.30.0
sentence-transformers==3.4.1
setuptools==77.0.3
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.6
SQLAlchemy==2.0.39
sqlparse==0.5.3
statsmodels==0.14.4
striprtf==0.0.26
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
trio==0.29.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
